updated CSVFileProcess
