+++ examples/sputterDeposition/config3D.txt
@@ -5,9 +5,9 @@ yExtent = 15
 
 offsetX = 12.
 offsetY = 15.
-holeRadius = 5.0
+holeRadius = 2.0
 holeDepth = 7.5
-taperingAngle = 7.5
+taperingAngle = 1.5
 
 # Process
 depositionTime = 1.5
